Still Linear MORE Torture
I finally got it to work by using the old code from more_linear_interpolation.py . The issue for me was quite straightforward and I just was not defining the XYZ locations for P1 and P2. I was instead using integers which was messing everything up, such an easy fix.